Remove cross-context referencing of IRC client.
It is safer to use IPC messages to communicate with the IRC client
across contexts. Cross-context referencing can cause strange invariant
violations with React and it is unsafe to pass callbacks to event
emitters from the renderer to the browser due to possible leaks if they
are not explicitly removed before the browser window reloads.